PR Review comments fix
